% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld01}
\alias{Hhld01}
\title{Table 01: Households by ownership type}
\format{
A tibble with 8 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code (0 = district total).}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{category}{Character. Housing ownership type (e.g., "Own",
    "Rented", "Institutional", "Other").}
  \item{number}{Character. Number of households.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld01
}
\description{
Number of households by type of ownership of housing unit, NPHC 2021.
}
\examples{
data(Hhld01)
head(Hhld01)
}
\keyword{datasets}
