% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld02}
\alias{Hhld02}
\title{Table 02: Households by foundation type}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (foundation material type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld02
}
\description{
Number of households by type of foundation of housing unit, NPHC 2021.
}
\examples{
data(Hhld02)
head(Hhld02)
}
\keyword{datasets}
