% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld03}
\alias{Hhld03}
\title{Table 03: Households by outer wall material}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (outer wall material), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld03
}
\description{
Number of households by type of materials used for outer wall of
housing unit, NPHC 2021.
}
\examples{
data(Hhld03)
head(Hhld03)
}
\keyword{datasets}
