% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld04}
\alias{Hhld04}
\title{Table 04: Households by roof material}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (roof material type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld04
}
\description{
Number of households by type of roof of housing unit, NPHC 2021.
}
\examples{
data(Hhld04)
head(Hhld04)
}
\keyword{datasets}
