% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld05}
\alias{Hhld05}
\title{Table 05: Households by floor material}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (floor material type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld05
}
\description{
Number of households by type of floor of housing unit, NPHC 2021.
}
\examples{
data(Hhld05)
head(Hhld05)
}
\keyword{datasets}
