% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld06}
\alias{Hhld06}
\title{Table 06: Households by drinking water source}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (water source type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld06
}
\description{
Number of households by type of main source of drinking water, NPHC 2021.
}
\examples{
data(Hhld06)
head(Hhld06)
}
\keyword{datasets}
