% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld07}
\alias{Hhld07}
\title{Table 07: Households by cooking fuel type}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (cooking fuel type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld07
}
\description{
Number of households by type of fuel usually used for cooking, NPHC 2021.
}
\examples{
data(Hhld07)
head(Hhld07)
}
\keyword{datasets}
