% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld08}
\alias{Hhld08}
\title{Table 08: Households by lighting source}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (lighting source), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld08
}
\description{
Number of households by usual source of lighting, NPHC 2021.
}
\examples{
data(Hhld08)
head(Hhld08)
}
\keyword{datasets}
