% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld09}
\alias{Hhld09}
\title{Table 09: Households by toilet type}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (toilet facility type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld09
}
\description{
Number of households by type of toilet facility, NPHC 2021.
}
\examples{
data(Hhld09)
head(Hhld09)
}
\keyword{datasets}
