% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld10}
\alias{Hhld10}
\title{Table 10: Households by amenities and assets}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (amenity/asset type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld10
}
\description{
Number of households by household amenities/assets, NPHC 2021.
}
\examples{
data(Hhld10)
head(Hhld10)
}
\keyword{datasets}
