% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld11}
\alias{Hhld11}
\title{Table 11: Households by female ownership of fixed assets}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (asset ownership category), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld11
}
\description{
Number of households by female household member's ownership of fixed
assets, NPHC 2021.
}
\examples{
data(Hhld11)
head(Hhld11)
}
\keyword{datasets}
