% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld12}
\alias{Hhld12}
\title{Table 12: Households with small-scale enterprises}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (enterprise type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld12
}
\description{
Number of households with small-scale enterprises, NPHC 2021.
}
\examples{
data(Hhld12)
head(Hhld12)
}
\keyword{datasets}
