% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld13}
\alias{Hhld13}
\title{Table 13: Households with at least one death}
\format{
A tibble with 10 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{tot_hhld}{Character. Total number of households.}
  \item{dead_hhld}{Character. Households with at least one death.}
  \item{category}{Character. Category label.}
  \item{number}{Character. Count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld13
}
\description{
Number of households where at least one person died in the 12 months
preceding the census, NPHC 2021.
}
\examples{
data(Hhld13)
head(Hhld13)
}
\keyword{datasets}
