% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld14}
\alias{Hhld14}
\title{Table 14: Deaths by age group}
\format{
A tibble with 10 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{agegrp}{Character. Age group code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{agegrpname}{Character. Age group label.}
  \item{category}{Character. Category label.}
  \item{number}{Character. Number of deaths.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld14
}
\description{
Number of deaths by age group of the deceased person, NPHC 2021.
}
\examples{
data(Hhld14)
head(Hhld14)
}
\keyword{datasets}
