% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld15}
\alias{Hhld15}
\title{Table 15: Deaths by age group, sex, and cause}
\format{
A tibble with 12 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{sex}{Character. Sex code (0 = total, 1 = male, 2 = female).}
  \item{agegrp}{Character. Age group code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{sexname}{Character. Sex label.}
  \item{agegrpname}{Character. Age group label.}
  \item{category}{Character. Cause of death category.}
  \item{number}{Character. Number of deaths.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld15
}
\description{
Number of deaths by age group of the deceased person, sex, and cause
of death, NPHC 2021.
}
\examples{
data(Hhld15)
head(Hhld15)
}
\keyword{datasets}
