% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld16}
\alias{Hhld16}
\title{Table 16: Female deaths by age group and status}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{agegrpname}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld16
}
\description{
Number of female deaths by broad age group of the deceased person and
status (e.g., maternal death), NPHC 2021.
}
\examples{
data(Hhld16)
head(Hhld16)
}
\keyword{datasets}
