% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld17}
\alias{Hhld17}
\title{Table 17: Households with absent members abroad}
\format{
A tibble with 10 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{tot_hhld}{Character. Total households.}
  \item{absnt_hhld}{Character. Households with at least one absent member.}
  \item{category}{Character. Category label.}
  \item{number}{Character. Count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld17
}
\description{
Number of households with at least one absentee member currently
abroad, NPHC 2021.
}
\examples{
data(Hhld17)
head(Hhld17)
}
\keyword{datasets}
