% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld18}
\alias{Hhld18}
\title{Table 18: Absent population abroad by sex}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld18
}
\description{
Absent population abroad by five-year age group and sex, NPHC 2021.
}
\examples{
data(Hhld18)
head(Hhld18)
}
\keyword{datasets}
