% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld19}
\alias{Hhld19}
\title{Table 19: Absent population abroad by age group and country}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname}, \code{category},
  \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld19
}
\description{
Absent population abroad by age group and country of destination,
NPHC 2021.
}
\examples{
data(Hhld19)
head(Hhld19)
}
\keyword{datasets}
