% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld20}
\alias{Hhld20}
\title{Table 20: Absent population abroad by age group and reason}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname}, \code{category},
  \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld20
}
\description{
Absent population abroad by age groups and reason of absence, NPHC 2021.
}
\examples{
data(Hhld20)
head(Hhld20)
}
\keyword{datasets}
