% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld21}
\alias{Hhld21}
\title{Table 21: Absent population abroad by country and reason}
\format{
A tibble with 12 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{sex}{Character. Sex code.}
  \item{reason}{Character. Reason of absence code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{sexname}{Character. Sex label.}
  \item{reasonname}{Character. Reason of absence label.}
  \item{category}{Character. Country of destination.}
  \item{number}{Character. Count of absent persons.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld21
}
\description{
Absent population abroad by country of destination and reason of
absence, NPHC 2021.
}
\examples{
data(Hhld21)
head(Hhld21)
}
\keyword{datasets}
