% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld22}
\alias{Hhld22}
\title{Table 22: Absent population abroad by duration and age at departure}
\format{
A tibble with 12 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{sex}{Character. Sex code.}
  \item{duratn}{Character. Duration of absence code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{sexname}{Character. Sex label.}
  \item{duratnname}{Character. Duration of absence label.}
  \item{category}{Character. Age at departure category.}
  \item{number}{Character. Count of absent persons.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld22
}
\description{
Absent population abroad by duration of absence and age at departure,
NPHC 2021.
}
\examples{
data(Hhld22)
head(Hhld22)
}
\keyword{datasets}
