% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hhld.R
\docType{data}
\name{Hhld23}
\alias{Hhld23}
\title{Table 23: Absent population abroad by age group and education level}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname}, \code{category}
  (education level), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Hhld23
}
\description{
Absent population abroad by age groups and level of education,
NPHC 2021.
}
\examples{
data(Hhld23)
head(Hhld23)
}
\keyword{datasets}
