% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv01}
\alias{Indv01}
\title{Table 01: Households and population by sex}
\format{
A tibble with 9 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{n_hhld}{Character. Number of households.}
  \item{category}{Character. Sex category ("Total", "Male", "Female").}
  \item{number}{Character. Population count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv01
}
\description{
Number of households and population by sex, NPHC 2021.
}
\examples{
data(Indv01)
head(Indv01)
}
\keyword{datasets}
