% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv02}
\alias{Indv02}
\title{Table 02: Households by household size}
\format{
A tibble with 8 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{provname}, \code{dname}, \code{gapaname},
  \code{category} (household size), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv02
}
\description{
Number of households by household size, NPHC 2021.
}
\examples{
data(Indv02)
head(Indv02)
}
\keyword{datasets}
