% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv03}
\alias{Indv03}
\title{Table 03: Population by single year of age and sex}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{agegrp} (single year code), \code{provname},
  \code{dname}, \code{gapaname}, \code{agegrpname} (age in years),
  \code{category} (sex), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv03
}
\description{
Population by single year of age and sex, NPHC 2021.
}
\examples{
data(Indv03)
head(Indv03)
}
\keyword{datasets}
