% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv05}
\alias{Indv05}
\title{Table 05: Local levels, households and population by locality size}
\format{
A tibble with 10 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{sizeclass}{Character. Locality size class code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{sizeclassname}{Character. Locality size class label.}
  \item{num_locality}{Character. Number of localities.}
  \item{no_hhld}{Character. Number of households.}
  \item{category}{Character. Sex category.}
  \item{number}{Character. Population count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv05
}
\description{
Number of local levels, households and population by size of locality,
NPHC 2021.
}
\examples{
data(Indv05)
head(Indv05)
}
\keyword{datasets}
