% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv06}
\alias{Indv06}
\title{Table 06: Household heads by age and sex}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{category} (age group),
  \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv06
}
\description{
Number of household heads by age group and sex of the head, NPHC 2021.
}
\examples{
data(Indv06)
head(Indv06)
}
\keyword{datasets}
