% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv07}
\alias{Indv07}
\title{Table 07: Population by relationship to household head}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{category} (relationship type),
  \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv07
}
\description{
Number of household members by relationship to household head and sex,
NPHC 2021.
}
\examples{
data(Indv07)
head(Indv07)
}
\keyword{datasets}
