% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv09}
\alias{Indv09}
\title{Table 09: Population by age group and nationality}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (nationality), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv09
}
\description{
Population by 5-year age group, sex, and country/region of nationality,
NPHC 2021.
}
\examples{
data(Indv09)
head(Indv09)
}
\keyword{datasets}
