% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv11}
\alias{Indv11}
\title{Table 11: Married population aged 10+ by age at first marriage}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (age at first marriage), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv11
}
\description{
Married population aged 10 years and above by 5-year age group, sex,
and age at first marriage, NPHC 2021.
}
\examples{
data(Indv11)
head(Indv11)
}
\keyword{datasets}
