% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv16}
\alias{Indv16}
\title{Table 16: Population by age group and disability type}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (disability type), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv16
}
\description{
Population by 5-year age group, sex, and type of disability, NPHC 2021.
}
\examples{
data(Indv16)
head(Indv16)
}
\keyword{datasets}
