% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv18}
\alias{Indv18}
\title{Table 18: Population aged 5+ by educational level}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (education level), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv18
}
\description{
Population aged 5 years and above by educational level attained and
sex, NPHC 2021.
}
\examples{
data(Indv18)
head(Indv18)
}
\keyword{datasets}
