% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv20}
\alias{Indv20}
\title{Table 20: Population aged 5-25 by current school attendance status}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{agegrpname}, \code{category} (attendance
  status), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv20
}
\description{
Population aged 5-25 years by status of school/college attendance,
NPHC 2021.
}
\examples{
data(Indv20)
head(Indv20)
}
\keyword{datasets}
