% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv21}
\alias{Indv21}
\title{Table 21: Population aged 15+ with 10+2 or higher by field of education}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (field of education), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv21
}
\description{
Population aged 15 years and above with 10+2 or above educational
level by field of education, NPHC 2021.
}
\examples{
data(Indv21)
head(Indv21)
}
\keyword{datasets}
