% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv22}
\alias{Indv22}
\title{Table 22: Population by place of birth}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (place of birth), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv22
}
\description{
Population by place of birth, sex, and 5-year age group, NPHC 2021.
}
\examples{
data(Indv22)
head(Indv22)
}
\keyword{datasets}
