% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv23}
\alias{Indv23}
\title{Table 23: Native-born population by province of birth}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (province of birth), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv23
}
\description{
Native-born population by province of birth, sex, and 5-year age
group, NPHC 2021.
}
\examples{
data(Indv23)
head(Indv23)
}
\keyword{datasets}
