% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv24}
\alias{Indv24}
\title{Table 24: Foreign-born population by country of birth}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (country of birth), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv24
}
\description{
Foreign-born population by broad age group and country of birth,
NPHC 2021.
}
\examples{
data(Indv24)
head(Indv24)
}
\keyword{datasets}
