% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv25}
\alias{Indv25}
\title{Table 25: Population by former place of residence}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (former place of residence), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv25
}
\description{
Population residing in enumerated area by former place of residence
(internal migration), NPHC 2021.
}
\examples{
data(Indv25)
head(Indv25)
}
\keyword{datasets}
