% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv26}
\alias{Indv26}
\title{Table 26: Migrated population by ecological belt}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (ecological belt), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv26
}
\description{
Population residing in enumerated area by former ecological belt of
residence, NPHC 2021.
}
\examples{
data(Indv26)
head(Indv26)
}
\keyword{datasets}
