% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv27}
\alias{Indv27}
\title{Table 27: Migrated population by former province of residence}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (former province), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv27
}
\description{
Population residing in enumerated area by former province of
residence, NPHC 2021.
}
\examples{
data(Indv27)
head(Indv27)
}
\keyword{datasets}
