% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv28}
\alias{Indv28}
\title{Table 28: Currently migrated population by length of stay}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (length of stay), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv28
}
\description{
Currently migrated population by length of stay and 5-year age group,
NPHC 2021.
}
\examples{
data(Indv28)
head(Indv28)
}
\keyword{datasets}
