% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv29}
\alias{Indv29}
\title{Table 29: Population by reason for migration (preceding census)}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (reason for migration), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv29
}
\description{
Population whose former place of residence preceding the census is
different from current, by reason for migration, NPHC 2021.
}
\examples{
data(Indv29)
head(Indv29)
}
\keyword{datasets}
