% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv30}
\alias{Indv30}
\title{Table 30: Population by former country of stay}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (country of stay), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv30
}
\description{
Population whose former place of residence is a foreign country, by
country of stay, NPHC 2021.
}
\examples{
data(Indv30)
head(Indv30)
}
\keyword{datasets}
