% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv31}
\alias{Indv31}
\title{Table 31: Migrated population by length of stay (5-year)}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (length of stay), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv31
}
\description{
Population whose former place of residence preceding the census is
different, by length of stay, NPHC 2021.
}
\examples{
data(Indv31)
head(Indv31)
}
\keyword{datasets}
