% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv32}
\alias{Indv32}
\title{Table 32: Migrated population by reason of stay}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (reason of stay), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv32
}
\description{
Population whose former place of residence preceding the census is
different, by reason of stay, NPHC 2021.
}
\examples{
data(Indv32)
head(Indv32)
}
\keyword{datasets}
