% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv33}
\alias{Indv33}
\title{Table 33: Number of children ever born}
\format{
A tibble with 10 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{nceb}{Character. Number of children ever born code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{childreneverborn}{Character. Number of children ever born label.}
  \item{category}{Character. Age group or other category.}
  \item{number}{Character. Count of women.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv33
}
\description{
Ever-married female population 15-49 years of age by number of
children ever born, NPHC 2021.
}
\examples{
data(Indv33)
head(Indv33)
}
\keyword{datasets}
