% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv34}
\alias{Indv34}
\title{Table 34: Number of children dead}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{ndead} (children dead code), \code{provname},
  \code{dname}, \code{gapaname}, \code{childrendead} (label),
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv34
}
\description{
Ever-married female population 15-49 years of age by number of
children who have died, NPHC 2021.
}
\examples{
data(Indv34)
head(Indv34)
}
\keyword{datasets}
