% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv35}
\alias{Indv35}
\title{Table 35: Number of children born alive in the last 12 months}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{code} (children born code), \code{provname},
  \code{dname}, \code{gapaname}, \code{codename} (label),
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv35
}
\description{
Ever-married female population 15-49 years of age by number of
children born alive in the last 12 months, NPHC 2021.
}
\examples{
data(Indv35)
head(Indv35)
}
\keyword{datasets}
