% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv36}
\alias{Indv36}
\title{Table 36: Children ever born, still alive, and died}
\format{
A tibble with 12 variables:
\describe{
  \item{prov}{Character. Province code.}
  \item{dist}{Character. District code.}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{chldtype}{Character. Child count type code (born/alive/dead).}
  \item{chldsex}{Character. Child sex code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{typename}{Character. Child count type label.}
  \item{sexname}{Character. Sex label.}
  \item{category}{Character. Age group of mother.}
  \item{number}{Character. Count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv36
}
\description{
Number of children ever born, still alive and died by sex of child,
NPHC 2021.
}
\examples{
data(Indv36)
head(Indv36)
}
\keyword{datasets}
