% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv37}
\alias{Indv37}
\title{Table 37: Children ever born by literacy and education of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{litsts} (literacy status code), \code{edulvl}
  (education level code), \code{provname}, \code{dname}, \code{gapaname},
  \code{educationlevel} (label), \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv37
}
\description{
Number of children ever born by literacy/educational status of mother,
NPHC 2021.
}
\examples{
data(Indv37)
head(Indv37)
}
\keyword{datasets}
