% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv38}
\alias{Indv38}
\title{Table 38: Children died by literacy and education of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{litsts}, \code{edulvl}, \code{provname},
  \code{dname}, \code{gapaname}, \code{educationlevel},
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv38
}
\description{
Number of children died by literacy/educational status of mother,
NPHC 2021.
}
\examples{
data(Indv38)
head(Indv38)
}
\keyword{datasets}
