% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv39}
\alias{Indv39}
\title{Table 39: Children born alive (last 12 months) by education of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{litsts}, \code{edulvl}, \code{provname},
  \code{dname}, \code{gapaname}, \code{educationlevel},
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv39
}
\description{
Number of children born alive in the last 12 months by literacy/
educational status of mother, NPHC 2021.
}
\examples{
data(Indv39)
head(Indv39)
}
\keyword{datasets}
