% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv40}
\alias{Indv40}
\title{Table 40: Children ever born by occupation of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{occ1} (1-digit occupation code), \code{provname},
  \code{dname}, \code{gapaname}, \code{occname} (occupation label),
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv40
}
\description{
Ever-married female population 15-49 years of age by major occupation
and number of children ever born, NPHC 2021.
}
\examples{
data(Indv40)
head(Indv40)
}
\keyword{datasets}
