% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv41}
\alias{Indv41}
\title{Table 41: Children died by occupation of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{occ1}, \code{provname}, \code{dname},
  \code{gapaname}, \code{occname}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv41
}
\description{
Number of children died by major occupation of mother, NPHC 2021.
}
\examples{
data(Indv41)
head(Indv41)
}
\keyword{datasets}
