% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv42}
\alias{Indv42}
\title{Table 42: Children born alive (last 12 months) by occupation of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{occ1}, \code{provname}, \code{dname},
  \code{gapaname}, \code{occname}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv42
}
\description{
Number of children born alive in last 12 months by major occupation of
mother, NPHC 2021.
}
\examples{
data(Indv42)
head(Indv42)
}
\keyword{datasets}
