% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv43}
\alias{Indv43}
\title{Table 43: Children ever born by industry of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{ind1} (1-digit industry code), \code{provname},
  \code{dname}, \code{gapaname}, \code{indname} (industry label),
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv43
}
\description{
Number of children ever born by major industry of mother, NPHC 2021.
}
\examples{
data(Indv43)
head(Indv43)
}
\keyword{datasets}
