% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv44}
\alias{Indv44}
\title{Table 44: Children died by industry of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{ind1}, \code{provname}, \code{dname},
  \code{gapaname}, \code{indname}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv44
}
\description{
Number of children died by major industry of mother, NPHC 2021.
}
\examples{
data(Indv44)
head(Indv44)
}
\keyword{datasets}
