% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv45}
\alias{Indv45}
\title{Table 45: Children born alive (last 12 months) by industry of mother}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{ind1}, \code{provname}, \code{dname},
  \code{gapaname}, \code{indname}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv45
}
\description{
Number of children born alive in last 12 months by major industry of
mother, NPHC 2021.
}
\examples{
data(Indv45)
head(Indv45)
}
\keyword{datasets}
