% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv46}
\alias{Indv46}
\title{Table 46: Children ever born by reason mother is not economically active}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{q36rsn} (reason code), \code{provname},
  \code{dname}, \code{gapaname}, \code{reasonnotactive} (label),
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv46
}
\description{
Number of children born alive in last 12 months by reason of economic
inactivity of mother, NPHC 2021.
}
\examples{
data(Indv46)
head(Indv46)
}
\keyword{datasets}
