% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv47}
\alias{Indv47}
\title{Table 47: Children dead by reason mother is not economically active}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{q36rsn}, \code{provname}, \code{dname},
  \code{gapaname}, \code{reasonnotactive}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv47
}
\description{
Number of children dead by reason of economic inactivity of mother,
NPHC 2021.
}
\examples{
data(Indv47)
head(Indv47)
}
\keyword{datasets}
