% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv48}
\alias{Indv48}
\title{Table 48: Children born alive (last 12 months) by reason mother not active}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{q36rsn}, \code{provname}, \code{dname},
  \code{gapaname}, \code{reasonnotactive}, \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv48
}
\description{
Number of children born alive in last 12 months by reason of economic
inactivity of mother, NPHC 2021.
}
\examples{
data(Indv48)
head(Indv48)
}
\keyword{datasets}
