% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv49}
\alias{Indv49}
\title{Table 49: Population aged 10+ by number of work months}
\format{
A tibble with 12 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{sex}{Character. Sex code.}
  \item{wrkmon}{Character. Work months code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{sexname}{Character. Sex label.}
  \item{workmonth}{Character. Work months label.}
  \item{category}{Character. Economic activity category.}
  \item{number}{Character. Population count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv49
}
\description{
Population 10 years of age and above by number of months worked in
the last 12 months, NPHC 2021.
}
\examples{
data(Indv49)
head(Indv49)
}
\keyword{datasets}
