% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv50}
\alias{Indv50}
\title{Table 50: Population aged 10+ by work months and occupation}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{wrkmon}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{workmonth},
  \code{category} (occupation), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv50
}
\description{
Population 10 years of age and above by number of months worked and
major occupation, NPHC 2021.
}
\examples{
data(Indv50)
head(Indv50)
}
\keyword{datasets}
