% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv51}
\alias{Indv51}
\title{Table 51: Population aged 10+ by work months and industry}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{wrkmon}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{workmonth},
  \code{category} (industry), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv51
}
\description{
Population 10 years of age and above by number of months worked and
major industry, NPHC 2021.
}
\examples{
data(Indv51)
head(Indv51)
}
\keyword{datasets}
