% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv53}
\alias{Indv53}
\title{Table 53: Economically active and inactive population aged 10+}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{agegrpname},
  \code{category} (economic activity status), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv53
}
\description{
Economically active and not active population 10 years of age and
above by age group, NPHC 2021.
}
\examples{
data(Indv53)
head(Indv53)
}
\keyword{datasets}
