% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv54}
\alias{Indv54}
\title{Table 54: Population aged 10+ by literacy and economic activity}
\format{
A tibble with 10 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{litsts}, \code{edulvl}, \code{provname},
  \code{dname}, \code{gapaname}, \code{educationlevel},
  \code{category}, \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv54
}
\description{
Population aged 10 years and above by literacy status and economic
activity, NPHC 2021.
}
\examples{
data(Indv54)
head(Indv54)
}
\keyword{datasets}
