% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv56}
\alias{Indv56}
\title{Table 56: Economically active population aged 10+ by industry}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{ind1}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{indname},
  \code{category} (age group), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv56
}
\description{
Economically active population 10 years of age and above by sex and
major industry, NPHC 2021.
}
\examples{
data(Indv56)
head(Indv56)
}
\keyword{datasets}
