% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv65}
\alias{Indv65}
\title{Table 65: Economically active population by industry and sector}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{ind1}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{indname},
  \code{category} (sector), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv65
}
\description{
Economically active population 10 years of age and above by sex,
major industry, and sector, NPHC 2021.
}
\examples{
data(Indv65)
head(Indv65)
}
\keyword{datasets}
