% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv66}
\alias{Indv66}
\title{Table 66: Population aged 10+ who did not work, by reason and age group}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (reason for not working), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv66
}
\description{
Population aged 10 years and above who did not work in the last 12
months by reason and age group, NPHC 2021.
}
\examples{
data(Indv66)
head(Indv66)
}
\keyword{datasets}
