% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv67}
\alias{Indv67}
\title{Table 67: Population aged 10+ who didn't work, by literacy and reason}
\format{
A tibble with 13 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{litsts}, \code{edulvl}, \code{provname},
  \code{dname}, \code{gapaname}, \code{sexname}, \code{educationlevel},
  \code{category} (reason), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv67
}
\description{
Population aged 10 years and above who didn't work in the last 12
months by sex, literacy status, and reason, NPHC 2021.
}
\examples{
data(Indv67)
head(Indv67)
}
\keyword{datasets}
