% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv68}
\alias{Indv68}
\title{Table 68: Population aged 10+ who did not work, by work months and reason}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{wrkmon} (work months code),
  \code{provname}, \code{dname}, \code{gapaname}, \code{sexname},
  \code{workmonth} (work months label), \code{category} (reason),
  \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv68
}
\description{
Population 10 years of age and over who did not work or worked less
than 12 months, by number of work months and reason, NPHC 2021.
}
\examples{
data(Indv68)
head(Indv68)
}
\keyword{datasets}
