% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv69}
\alias{Indv69}
\title{Table 69: Population under 18 by living arrangement (sex and age)}
\format{
A tibble with 12 variables: \code{prov}, \code{dist},
  \code{gapa}, \code{sex}, \code{agegrp}, \code{provname}, \code{dname},
  \code{gapaname}, \code{sexname}, \code{agegrpname},
  \code{category} (living arrangement), \code{number}.
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv69
}
\description{
Population under 18 years of age by status of living arrangement,
sex, and age group, NPHC 2021.
}
\examples{
data(Indv69)
head(Indv69)
}
\keyword{datasets}
