% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Indv.R
\docType{data}
\name{Indv70}
\alias{Indv70}
\title{Table 70: Population under 18 by living arrangement and literacy status}
\format{
A tibble with 12 variables:
\describe{
  \item{prov}{Character. Province code (0 = Nepal total).}
  \item{dist}{Character. District code (0 = province total).}
  \item{gapa}{Character. Gaunpalika/Nagarpalika code.}
  \item{sex}{Character. Sex code.}
  \item{litsts}{Character. Literacy status code.}
  \item{provname}{Character. Province name.}
  \item{dname}{Character. District name.}
  \item{gapaname}{Character. Gaunpalika/Nagarpalika name.}
  \item{sexname}{Character. Sex label.}
  \item{literacystatus}{Character. Literacy status label.}
  \item{category}{Character. Living arrangement category.}
  \item{number}{Character. Population count.}
}
}
\source{
National Statistics Office (NSO) of Nepal, NPHC 2021.
}
\usage{
Indv70
}
\description{
Population under 18 years of age by status of living arrangement and
literacy status, NPHC 2021.
}
\examples{
data(Indv70)
head(Indv70)
}
\keyword{datasets}
