% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepalcensus-package.R
\docType{package}
\name{nepalcensus}
\alias{nepalcensus}
\alias{nepalcensus-package}
\title{nepalcensus: Nepal Population and Housing Census 2021 Data}
\source{
National Statistics Office (NSO) of Nepal.
  \url{https://censusnepal.cbs.gov.np}
}
\description{
Provides tidy access to 89 tabular datasets from the Nepal Population
and Housing Census 2021 (NPHC 2078 BS). Datasets cover household
characteristics and individual-level demographics cross-tabulated by
province, district, and local level.
}
\section{Household datasets (Hhld01--Hhld23)}{

Housing ownership, construction materials, utilities, sanitation,
household amenities, death records, and absent population abroad.
}

\section{Individual datasets (Indv01--Indv71)}{

Population by sex, age, nationality, marital status, disability,
literacy, education, migration, fertility, economic activity,
occupation, industry, employment status, and birth registration.
}

\section{Helper functions}{

\describe{
  \item{\code{\link{catalog}}}{Browse or search the data catalog.}
  \item{\code{\link{search_census}}}{Keyword search across dataset titles.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bjungbogati/nepalcensus}
  \item Report bugs at \url{https://github.com/bjungbogati/nepalcensus/issues}
}

}
\author{
\strong{Maintainer}: Binoj Jungbogati \email{binodb@dizzaroo.com}

Authors:
\itemize{
  \item Binoj Jungbogati \email{binodb@dizzaroo.com}
}

}
